requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
django-q2==1.9.0
